refactor(orchestration): move network fakes from test to tools
