+++ packages/orchestration/test/examples/stake-ica.contract.test.ts
@@ -9,7 +9,7 @@ import { type StakeIcaTerms } from '../../src/examples/stake-ica.contract.js';
 import fetchedChainInfo from '../../src/fetched-chain-info.js';
 import type { DenomAmount } from '../../src/orchestration-api.js';
 import { maxClockSkew } from '../../src/utils/cosmos.js';
-import { UNBOND_PERIOD_SECONDS } from '../ibc-mocks.js';
+import { UNBOND_PERIOD_SECONDS } from '../../tools/ibc-mock-fixtures.js';
 import { commonSetup } from '../supports.js';
 
 import * as contractExports from '../../src/examples/stake-ica.contract.js';
