+++ packages/orchestration/test/examples/staking-combinations.test.ts
@@ -15,7 +15,10 @@ import {
   buildMsgResponseString,
   parseOutgoingTxPacket,
 } from '../../tools/ibc-mocks.js';
-import { protoMsgMocks, UNBOND_PERIOD_SECONDS } from '../ibc-mocks.js';
+import {
+  protoMsgMocks,
+  UNBOND_PERIOD_SECONDS,
+} from '../../tools/ibc-mock-fixtures.js';
 import { commonSetup } from '../supports.js';
 
 type StartFn = typeof contractExports.start;
