+++ scripts/ci/tools-scope-policy.mjs
@@ -65,10 +65,6 @@ export const legacySrcToToolsImports = [
 
 export const legacyNonTestToTestImports = [
   {
-    file: 'packages/orchestration/tools/contract-tests.ts',
-    specifier: '@agoric/orchestration/test/network-fakes.js',
-  },
-  {
     file: 'packages/swingset-liveslots/tools/vo-test-harness.js',
     specifier: '../test/liveslots-helpers.js',
   },
